Function-level implementations - step 1
